+++ R/get_cov.R
@@ -1,44 +1,68 @@
 #' Extract covariates of choice occasion
 #'
 #' @description
-#' This convenience function returns the covariates and the choices of specific
+#' This helper function returns the covariates and the choices of specific
 #' choice occasions.
 #'
 #' @param x
 #' Either an object of class \code{RprobitB_data} or \code{RprobitB_fit}.
 #'
-#' @param id
-#' A numeric (vector), that specifies the decider(s).
+#' @param id,idc \[`integer()` | `NULL`\]\cr
+#' Identifiers for deciders and choice occasions.
 #'
-#' @param idc
-#' A numeric (vector), that specifies the choice occasion(s).
+#' If `NULL`, everything is returned.
 #'
-#' @param idc_label
-#' The name of the column that contains the choice occasion identifier.
+#' @param id_label,idc_label \[`character(1)` | `NULL`\]\cr
+#' The columns that contain the decider and choice occasion identifier.
+#'
+#' If `NULL`, this information is extracted from `x`.
 #'
 #' @return
 #' A subset of the `choice_data` data frame specified in `prepare_data()`.
 #'
 #' @export
+#'
+#' @examples
+#' data <- simulate_choices(
+#'   form = product ~ price,
+#'   N = 10,
+#'   T = 1:10,
+#'   J = 3,
+#'   ranked = TRUE
+#' )
+#' get_cov(data, id = 2)
 
-get_cov <- function(x, id, idc, idc_label) {
-  if (inherits(x, "RprobitB_fit")) {
-    x <- x$data
-  }
-  if (inherits(x, "RprobitB_data")) {
-    id_label <- x$res_var_names$id
-    idc_label <- if (missing(idc_label)) x$res_var_names$idc else idc_label
-    if (missing(id)) id <- x$choice_data[[id_label]]
-    if (missing(idc)) idc <- x$choice_data[[idc_label]]
-    ind <- x$choice_data[[id_label]] %in% id & x$choice_data[[idc_label]] %in% idc
-    out <- x$choice_data[ind, ]
-    if (nrow(out) == 0) {
-      stop("Requested choice occasion not found.", call. = FALSE)
-    }
-    return(out)
-  } else {
-    stop("'x' must be either an 'RprobitB_fit' or 'RprobitB_data' object.",
-         call. = FALSE
-    )
-  }
+get_cov <- function(
+    x, id = NULL, idc = NULL, id_label = NULL, idc_label = NULL
+  ) {
+  oeli::input_check_response(
+    check = list(
+      checkmate::check_class(x, "RprobitB_data"),
+      checkmate::check_class(x, "RprobitB_fit")
+    ),
+    var_name = "x"
+  )
+  if (inherits(x, "RprobitB_fit")) x <- x$data
+  id_label <- if (is.null(id_label)) x$res_var_names$id else id_label
+  idc_label <- if (is.null(idc_label)) x$res_var_names$idc else idc_label
+  oeli::input_check_response(
+    check = checkmate::check_string(id_label),
+    var_name = "id_label"
+  )
+  oeli::input_check_response(
+    check = checkmate::check_string(idc_label),
+    var_name = "idc_label"
+  )
+  if (is.null(id)) id <- x$choice_data[[id_label]]
+  if (is.null(idc)) idc <- x$choice_data[[idc_label]]
+  oeli::input_check_response(
+    check = checkmate::check_integerish(id, lower = 1, any.missing = FALSE),
+    var_name = "id"
+  )
+  oeli::input_check_response(
+    check = checkmate::check_integerish(idc, lower = 1, any.missing = FALSE),
+    var_name = "idc"
+  )
+  ind <- x$choice_data[[id_label]] %in% id & x$choice_data[[idc_label]] %in% idc
+  x$choice_data[ind, ]
 }